Add support to bundled and custom executables with version

+++ src/main/groovy/org/web3j/solidity/gradle/plugin/SolidityCompiler.groovy
@@ -12,8 +12,14 @@ import java.nio.file.StandardCopyOption
 class SolidityCompiler {
 
     private File solc = null
+    private ClassLoader classLoader
 
     SolidityCompiler() {
+        this(null)
+    }
+
+    SolidityCompiler(final ClassLoader classLoader) {
+        this.classLoader = classLoader
         try {
             init()
         } catch (IOException e) {
@@ -29,24 +35,34 @@ class SolidityCompiler {
         final def tmpDir = new File(System.getProperty('java.io.tmpdir'), 'solc')
         tmpDir.mkdirs()
 
-        final def binariesPath = "/native/${getSystemDirectory()}/solc"
-        final def stream = getClass().getResourceAsStream("$binariesPath/file.list")
+        final def binariesPath = "native/${getSystemDirectory()}/solc"
+        final def stream = getResourceAsStream("$binariesPath/file.list")
 
-        new Scanner(stream as InputStream).withCloseable {
-            while (it.hasNext()) {
-                final def scan = it.next()
-                final def targetFile = new File(tmpDir, scan)
-                targetFile.deleteOnExit()
+        if (stream != null) {
+            new Scanner(stream as InputStream).withCloseable {
+                while (it.hasNext()) {
+                    final def scan = it.next()
+                    final def targetFile = new File(tmpDir, scan)
+                    targetFile.deleteOnExit()
 
-                final def fis = getClass().getResourceAsStream("$binariesPath/$scan")
-                Files.copy(fis, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING)
+                    final def fis = getResourceAsStream("$binariesPath/$scan")
+                    Files.copy(fis, targetFile.toPath(), StandardCopyOption.REPLACE_EXISTING)
 
-                if (solc == null) {
-                    // first file in the list denotes executable
-                    solc = targetFile
-                    solc.setExecutable(true)
+                    if (solc == null) {
+                        // first file in the list denotes executable
+                        solc = targetFile
+                        solc.setExecutable(true)
+                    }
                 }
             }
+        }
+    }
+
+    private InputStream getResourceAsStream(final String name) {
+        if (classLoader != null) {
+            return classLoader.getResourceAsStream(name)
+        } else {
+            return getClass().getResourceAsStream("/$name")
         }
     }
